+++ tomcat/src/test/java/com/techcourse/servlet/LoginControllerTest.java
@@ -1,4 +1,4 @@
-package org.apache;
+package com.techcourse.servlet;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -16,6 +16,8 @@ import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import com.techcourse.servlet.LoginController;
 
 import support.HttpRequestFixture;
